Analytics/requirements.txt
paho-mqtt
requests
nats-py